implement type validation

+++ src/Webhook/TimeEntry/Timer.php
@@ -51,17 +51,19 @@ class Timer
 
     private function storeTimeEntryTagIds($workspaceId, $userId, $projectId, $timeEntryId, $tagIds)
     {
-        DB::table('time_entry_tags')->where('timeEntryId', $timeEntryId)->delete();
-        foreach ($tagIds as $tagId) {
-            $data['workspaceId'] = $workspaceId;
-            $data['userId'] = $userId;
-            $data['projectId'] = $projectId;
-            $data['timeEntryId'] = $timeEntryId;
-            $data['tagId'] = $tagId['id'];
-            DB::table('time_entry_tags')->updateOrInsert(
-                ['timeEntryId' => $timeEntryId, 'tagId' => $tagId['id']],
-                $data
-            );
+        if (is_array($tagIds) || is_object($tagIds)) {
+            DB::table('time_entry_tags')->where('timeEntryId', $timeEntryId)->delete();
+            foreach ($tagIds as $tagId) {
+                $data['workspaceId'] = $workspaceId;
+                $data['userId'] = $userId;
+                $data['projectId'] = $projectId;
+                $data['timeEntryId'] = $timeEntryId;
+                $data['tagId'] = $tagId['id'];
+                DB::table('time_entry_tags')->updateOrInsert(
+                    ['timeEntryId' => $timeEntryId, 'tagId' => $tagId['id']],
+                    $data
+                );
+            }
         }
     }
     private function storeTimeEntryInterval($workspaceId, $userId, $projectId, $timeEntryId, $timeInterval)
